# api/swagger/v1/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/training": {
      "get": {
        "summary": "Read all trainees",
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "description": "API versioning: specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TraineeService"
        ]
      },
      "post": {
        "summary": "Create new trainee",
        "operationId": "Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRequest"
            }
          }
        ],
        "tags": [
          "TraineeService"
        ]
      }
    },
...
  },
  "definitions": {
    "v1CreateRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: specify version explicitly"
        },
        "trainee": {
          "$ref": "#/definitions/v1Trainee",
          "title": "Task entity to add"
        }
      },
      "title": "Request data to create new trainee"
    },
    "v1CreateResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: specify version explicitly"
        },
        "id": {
          "type": "string",
          "format": "int64",
          "title": "ID of created trainee"
        }
      },
      "title": "Contains data of created trainee"
    },
...
    "v1ReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: specify version explicitly"
        },
        "trainees": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Trainee"
          },
          "title": "List of all trainees"
        }
      },
      "title": "Contains list of all trainees"
    },
...
    "v1Trainee": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Unique integer identifier"
        },
        "name": {
          "type": "string",
          "title": "Trainee first name"
        },
        "surname": {
          "type": "string",
          "title": "Trainee last name"
        },
        "email": {
          "type": "string",
          "title": "Trainee e-mail"
        },
        "hasAttend": {
          "type": "boolean",
          "format": "boolean",
          "title": "Trainee attendance"
        },
        "score": {
          "type": "integer",
          "format": "int32",
          "title": "Points number received for the task"
        }
      },
      "title": "Main task"
    },
...
  }
}